+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 0.9.0.0 (UNRELEASED, DEV)
 
+- API
+  - Dropped `path` and `path-io` in favour of good ol' `FilePath`
+    - This also lifts the restriction with absolute paths
 - Misc changes
   - #145: CLI arguments have been revamped
     - Added `--input-dir/--output-dir` to override these paths
     - Accept host string in addition to port number
     - Exposed `Rib.Shake.getCliConfig` to get full CLI configuration
+    - Allow customizing fsnotify ignore list
   - #141: Allow quiet logging (useful when rib is used as a library)
 
 ## 0.8.0.0